Apply review feedback: Remove unnecessary cast
Credit: Issue found during a Claude Code code review.

+++ packages/onenote-converter/parser-utils/src/reader.rs
@@ -185,7 +185,7 @@ impl<'a> Reader<'a> {
             .into());
         }
 
-        self.seek(SeekFrom::Current(count as i64))? as usize;
+        self.seek(SeekFrom::Current(count as i64))?;
 
         Ok(())
     }
